Fixed dat_core imports
