+++ verified_destinations/destination_pinecone/destination_pinecone.py
@@ -1,23 +1,24 @@
 import os
-from typing import Any, Tuple, Optional, Mapping, Iterable
-from connectors.destinations.destination import Destination
-from dat_core.pydantic_models.connector_specification import ConnectorSpecification
-from connectors.destinations.destination_qdrant.seeder import QdrantSeeder
-from connectors.destinations.vector_db_helpers.data_processor import DataProcessor
+from dat_core.connectors.destinations.destination import Destination
+from typing import Any, Iterable, Mapping, Tuple, Optional
+from dat_core.pydantic_models.connector_specification import ConnectorSpecification, DestinationSyncMode
 from dat_core.pydantic_models.dat_catalog import DatCatalog
 from dat_core.pydantic_models.dat_message import DatMessage, Type, DatDocumentMessage, Data
-from dat_core.pydantic_models.dat_document_stream import SyncMode, DatDocumentStream
+from verified_destinations.destination_pinecone.seeder import PineconeSeeder
+from dat_core.connectors.destinations.vector_db_helpers.data_processor import DataProcessor
+from dat_core.pydantic_models.dat_document_stream import DatDocumentStream, SyncMode
 
 
 BATCH_SIZE = 1000
 
-class Qdrant(Destination):
-    
+
+class Pinecone(Destination):
+
     _spec_file = os.path.join(os.path.dirname(
         os.path.abspath(__file__)), 'specs.yml')
 
     def _init_seeder(self, config: Mapping[str, Any]) -> None:
-        self.seeder = QdrantSeeder(config, config.connectionSpecification.get('embedding_dimensions'))
+        self.seeder = PineconeSeeder(config, config.connectionSpecification.get('embedding_dimensions'))
 
     def check_connection(self, config: ConnectorSpecification) -> Tuple[bool, Optional[Any]]:
         self._init_seeder(config)
@@ -32,25 +33,26 @@ class Qdrant(Destination):
         processor = DataProcessor(config, self.seeder, BATCH_SIZE)
         yield from processor.processor(configured_catalog, input_messages)
 
+
 if __name__ == '__main__':
     destination_config_file = os.path.join(os.path.dirname(
         os.path.abspath(__file__)), 'destination_config.json')
     config = ConnectorSpecification.model_validate_json(
             open(destination_config_file).read(), )
     configured_catalog = DatCatalog.model_validate_json(
-            open('./connectors/destinations/destination_qdrant/configured_catalog.json').read(), )
+            open('./verified_destinations/destination_pinecone/configured_catalog.json').read(), )
     print(config)
-    _spec = Qdrant().spec()
+    _spec = Pinecone().spec()
     print(f"spec: {_spec}")
-    _dest = Qdrant().check(config=config)
+    _dest = Pinecone().check(config=config)
     print(f"destination check: {_dest}")
-    docs = Qdrant().write(config=config,configured_catalog=configured_catalog , input_messages=[
+    docs = Pinecone().write(config=config,configured_catalog=configured_catalog , input_messages=[
         DatMessage(
             type=Type.RECORD,
             record=DatDocumentMessage(
                 data=Data(
                     document_chunk='foo',
-                    vectors=[0.0] * 1536,
+                    vectors=[0.1] * 1536,
                     metadata={"meta": "Objective", "dat_source": "S3", "dat_stream": "PDF", "dat_document_entity": "DBT/DBT Overview.pdf"},
                 ),
                 emitted_at=1,
@@ -66,7 +68,7 @@ if __name__ == '__main__':
             record=DatDocumentMessage(
                 data=Data(
                     document_chunk='bar',
-                    vectors=[0.0] * 1536,
+                    vectors=[1.0] * 1536,
                     metadata={"meta": "Arbitrary", "dat_source": "S3", "dat_stream": "PDF", "dat_document_entity": "DBT/DBT Overview.pdf"},
                 ),
                 emitted_at=2,
