+++ verified_destinations/destination_pinecone/specs.yml
