+++ verified_destinations/destination_pinecone/tests/test_pinecone.py
@@ -1,19 +1,17 @@
+import io
 import yaml
 from typing import List
 from dat_core.pydantic_models.connector_specification import ConnectorSpecification
 from dat_core.pydantic_models.dat_connection_status import Status
+from verified_destinations.destination_pinecone.destination_pinecone import Pinecone
 from conftest import *
-from connectors.destinations.destination_qdrant.destination_qdrant import Qdrant
 from dat_core.pydantic_models.dat_message import (DatMessage, DatDocumentMessage,
-                                         Data, DatStateMessage, StreamState,
-                                         StreamStatus, DatDocumentStream,
-                                         Type)
+                                         Data, DatStateMessage,
+                                         StreamState, StreamStatus,
+                                         DatDocumentStream, Type)
 from dat_core.pydantic_models.dat_catalog import DatCatalog
 
-
-class TestQdrant:
-
-    DESTINATION_CONFIG_FILE = "./connectors/destinations/destination_qdrant/destination_config.json"
+class TestPinecone:
 
     def test_spec(self, ):
         """
@@ -21,8 +19,8 @@ class TestQdrant:
         WHEN spec() is called on a valid Destination class
         THEN spec stated in ./specs/ConnectorSpecification.yml is returned
         """
-        spec = Qdrant().spec()
-        with open('./connectors/destinations/destination_qdrant/specs.yml') as yaml_in:
+        spec = Pinecone().spec()
+        with open('./verified_destinations/destination_pinecone/specs.yml') as yaml_in:
             schema = yaml.safe_load(yaml_in)
             assert schema == spec
 
@@ -32,29 +30,31 @@ class TestQdrant:
         WHEN check() is called on a valid Destination class
         THEN no error is raised
         """
+        destination_config_file = "./verified_destinations/destination_pinecone/destination_config.json"
         config = ConnectorSpecification.model_validate_json(
-            open(self.DESTINATION_CONFIG_FILE).read(), )
-        check = Qdrant().check(
+            open(destination_config_file).read(), )
+        check = Pinecone().check(
             config=config)
         print(check)
         assert check.status == Status.SUCCEEDED
-    
+
     def test_write(self, ):
         """
-        Given a valid connectionSpecification JSON config
+        GIVEN a valid connectionSpecification JSON config
         WHEN write() is called on a valid Destination class
         THEN no error is raised
         """
+        destination_config_file = "./verified_destinations/destination_pinecone/destination_config.json"
         config = ConnectorSpecification.model_validate_json(
-            open(self.DESTINATION_CONFIG_FILE).read(), )
+            open(destination_config_file).read(), )
         configured_catalog = DatCatalog.model_validate_json(
-            open('./connectors/destinations/destination_qdrant/configured_catalog.json').read(), )
+            open('./verified_destinations/destination_pinecone/configured_catalog.json').read(), )
         first_record = DatMessage(
                 type=Type.RECORD,
                 record=DatDocumentMessage(
                     data=Data(
                         document_chunk='foo',
-                        vectors=[0.0] * 1536,
+                        vectors=[0.1] * 1536,
                         metadata={"meta": "Objective", "dat_source": "S3",
                                   "dat_stream": "PDF", "dat_document_entity": "DBT/DBT Overview.pdf"},
                     ),
@@ -72,7 +72,7 @@ class TestQdrant:
                 record=DatDocumentMessage(
                     data=Data(
                         document_chunk='bar',
-                        vectors=[0.0] * 1536,
+                        vectors=[1.0] * 1536,
                         metadata={"meta": "Arbitrary", "dat_source": "S3",
                                   "dat_stream": "PDF", "dat_document_entity": "Apple/DBT/DBT Overview.pdf"},
                     ),
@@ -89,12 +89,7 @@ class TestQdrant:
             first_record,
             second_record,
         ]
-        check  = Qdrant().check(
-            config=config)
-        # import pdb;pdb.set_trace()
-        assert check.status == Status.SUCCEEDED
-
-        docs = Qdrant().write(
+        docs = Pinecone().write(
             config=config,
             configured_catalog=configured_catalog,
             input_messages=mocked_input
@@ -109,16 +104,17 @@ class TestQdrant:
         WHEN write() is called on a valid Destination class
         THEN no error is raised
         """
+        destination_config_file = "./verified_destinations/destination_pinecone/destination_config.json"
         config = ConnectorSpecification.model_validate_json(
-            open(self.DESTINATION_CONFIG_FILE).read(), )
+            open(destination_config_file).read(), )
         configured_catalog = DatCatalog.model_validate_json(
-            open('./connectors/destinations/destination_qdrant/configured_catalog.json').read(), )
+            open('./verified_destinations/destination_pinecone/configured_catalog.json').read(), )
         first_record = DatMessage(
                 type=Type.RECORD,
                 record=DatDocumentMessage(
                     data=Data(
                         document_chunk='foo',
-                        vectors=[0.0] * 1536,
+                        vectors=[1.0] * 1536,
                         metadata={"meta": "Objective", "dat_source": "S3",
                                   "dat_stream": "PDF", "dat_document_entity": "DBT/DBT Overview.pdf"},
                     ),
@@ -136,14 +132,14 @@ class TestQdrant:
                 record=DatDocumentMessage(
                     data=Data(
                         document_chunk='bar',
-                        vectors=[0.0] * 1536,
+                        vectors=[1.1] * 1536,
                         metadata={"meta": "Arbitrary", "dat_source": "S3",
                                   "dat_stream": "PDF", "dat_document_entity": "Apple/DBT/DBT Overview.pdf"},
                     ),
                     emitted_at=2,
                     namespace="pytest_seeder",
                     stream=DatDocumentStream(
-                        name="S3",
+                        name="GCS",
                         namespace="pytest_seeder",
                         sync_mode="incremental",
                     )
@@ -170,7 +166,7 @@ class TestQdrant:
                 type=Type.STATE,
                 state=DatStateMessage(
                     stream=DatDocumentStream(
-                        name="S3",
+                        name="GCS",
                         namespace="pytest_seeder",
                         sync_mode="incremental",
                     ),
@@ -197,17 +193,11 @@ class TestQdrant:
                 ),
             ),
         ]
-        check  = Qdrant().check(
-            config=config)
-        # import pdb;pdb.set_trace()
-        assert check.status == Status.SUCCEEDED
-
-        docs = Qdrant().write(
+        docs = Pinecone().write(
             config=config,
             configured_catalog=configured_catalog,
             input_messages=mocked_input
         )
-
         for doc in docs:
             print(f"doc: {doc}")
             assert isinstance(doc, DatMessage)
