+++ verified_destinations/destination_qdrant/destination_qdrant.py
@@ -1,25 +1,23 @@
 import os
-from connectors.destinations.destination import Destination
-from typing import Any, Iterable, Mapping, Tuple, Optional
-from dat_core.pydantic_models.connector_specification import ConnectorSpecification, DestinationSyncMode
+from typing import Any, Tuple, Optional, Mapping, Iterable
+from dat_core.connectors.destinations.destination import Destination
+from dat_core.pydantic_models.connector_specification import ConnectorSpecification
+from verified_destinations.destination_qdrant.seeder import QdrantSeeder
+from dat_core.connectors.destinations.vector_db_helpers.data_processor import DataProcessor
 from dat_core.pydantic_models.dat_catalog import DatCatalog
 from dat_core.pydantic_models.dat_message import DatMessage, Type, DatDocumentMessage, Data
-from connectors.destinations.destination_pinecone.seeder import PineconeSeeder
-from connectors.destinations.vector_db_helpers.data_processor import DataProcessor
-from dat_core.pydantic_models.dat_document_stream import DatDocumentStream, SyncMode
-
+from dat_core.pydantic_models.dat_document_stream import SyncMode, DatDocumentStream
 
 
 BATCH_SIZE = 1000
 
-
-class Pinecone(Destination):
-
+class Qdrant(Destination):
+    
     _spec_file = os.path.join(os.path.dirname(
         os.path.abspath(__file__)), 'specs.yml')
 
     def _init_seeder(self, config: Mapping[str, Any]) -> None:
-        self.seeder = PineconeSeeder(config, config.connectionSpecification.get('embedding_dimensions'))
+        self.seeder = QdrantSeeder(config, config.connectionSpecification.get('embedding_dimensions'))
 
     def check_connection(self, config: ConnectorSpecification) -> Tuple[bool, Optional[Any]]:
         self._init_seeder(config)
@@ -34,26 +32,25 @@ class Pinecone(Destination):
         processor = DataProcessor(config, self.seeder, BATCH_SIZE)
         yield from processor.processor(configured_catalog, input_messages)
 
-
 if __name__ == '__main__':
     destination_config_file = os.path.join(os.path.dirname(
         os.path.abspath(__file__)), 'destination_config.json')
     config = ConnectorSpecification.model_validate_json(
             open(destination_config_file).read(), )
     configured_catalog = DatCatalog.model_validate_json(
-            open('./connectors/destinations/destination_pinecone/configured_catalog.json').read(), )
+            open('./verified_destinations/destination_qdrant/configured_catalog.json').read(), )
     print(config)
-    _spec = Pinecone().spec()
+    _spec = Qdrant().spec()
     print(f"spec: {_spec}")
-    _dest = Pinecone().check(config=config)
+    _dest = Qdrant().check(config=config)
     print(f"destination check: {_dest}")
-    docs = Pinecone().write(config=config,configured_catalog=configured_catalog , input_messages=[
+    docs = Qdrant().write(config=config,configured_catalog=configured_catalog , input_messages=[
         DatMessage(
             type=Type.RECORD,
             record=DatDocumentMessage(
                 data=Data(
                     document_chunk='foo',
-                    vectors=[0.1] * 1536,
+                    vectors=[0.0] * 1536,
                     metadata={"meta": "Objective", "dat_source": "S3", "dat_stream": "PDF", "dat_document_entity": "DBT/DBT Overview.pdf"},
                 ),
                 emitted_at=1,
@@ -69,7 +66,7 @@ if __name__ == '__main__':
             record=DatDocumentMessage(
                 data=Data(
                     document_chunk='bar',
-                    vectors=[1.0] * 1536,
+                    vectors=[0.0] * 1536,
                     metadata={"meta": "Arbitrary", "dat_source": "S3", "dat_stream": "PDF", "dat_document_entity": "DBT/DBT Overview.pdf"},
                 ),
                 emitted_at=2,
