+++ verified_destinations/destination_qdrant/seeder.py
@@ -1,7 +1,7 @@
 import uuid
 from qdrant_client import QdrantClient, models
 from typing import Any, List, Optional, Dict
-from connectors.destinations.vector_db_helpers.seeder import Seeder
+from dat_core.connectors.destinations.vector_db_helpers.seeder import Seeder
 from qdrant_openapi_client.models import VectorParams, Distance
 from dat_core.pydantic_models.dat_message import DatDocumentMessage
 from dat_core.pydantic_models.stream_metadata import StreamMetadata
