+++ verified_destinations/destination_qdrant/specs.yml
